Fixed lot's of imports and started refactoring some of the basic Form Fields
