+++ packages/pydantic-forms/package.json
@@ -13,6 +13,7 @@
     },
     "dependencies": {
         "invariant": "^2.2.4",
+        "object-hash": "^2.1.1",
         "uniforms": "^3.8.1",
         "uniforms-bridge-json-schema": "^3.8.1",
         "uniforms-unstyled": "^3.8.1"
@@ -43,6 +44,7 @@
         "@testing-library/user-event": "^14.4.3",
         "@types/invariant": "^2.2.33",
         "@types/jest": "^27.4.0",
+        "@types/object-hash": "^2.1.0",
         "esbuild": "^0.14.10",
         "esbuild-jest": "^0.5.0",
         "jest": "^27.4.7",
